refactor: отделил middleware в отдельный модуль и подготовил файлы для новой функциональности через Middleware
