+++ src/pybot/bot/middlewares/__init__.py
@@ -1,0 +1,5 @@
+from .db import DbSessionMiddleware
+
+__all__ = [
+    "DbSessionMiddleware",
+]
