+++ src/pybot/bot/middlewares/db.py
@@ -1,0 +1,26 @@
+from collections.abc import Awaitable, Callable
+from typing import Any
+
+# простейший middleware для сессии БД
+from aiogram import BaseMiddleware
+from aiogram.types import TelegramObject, Update
+from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker
+
+
+class DbSessionMiddleware(BaseMiddleware):
+    def __init__(self, session_maker: async_sessionmaker[AsyncSession]) -> None:
+        super().__init__()
+        self.session_maker = session_maker
+
+    async def __call__(
+        self,
+        handler: Callable[[TelegramObject, dict[str, Any]], Awaitable[Any]],
+        event: TelegramObject,
+        data: dict[str, Any],
+    ) -> Any:
+        if not isinstance(event, Update):
+            return await handler(event, data)
+
+        async with self.session_maker() as session:
+            data["db"] = session
+            return await handler(event, data)
